Fix: RecipeDetailFetchService삭제후 Repository에 Service를 적용
